v3/requirements.txt
google-cloud-datastore==1.5.0
google-cloud-storage==1.8.0
requests
argparse